pyproject.toml
[tool.poetry]
name = "app-estilo"
version = "0.1.0"
description = "API RESTful para Estilo"
authors = ["Desenvolvedor <dev@example.com>"]
packages = [
    { include = "app" }
]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.12"
uvicorn = {extras = ["standard"], version = "^0.34.2"}
sqlalchemy = "^2.0.41"
alembic = "^1.16.1"
python-jose = {extras = ["cryptography"], version = "^3.4.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.20"
psycopg2-binary = "^2.9.10"
pydantic = "^2.6.4"
pydantic-settings = "^2.2.1"
email-validator = "^2.1.0.post1"
pytest-cov = "^6.1.1"
bcrypt = ">=3.2.0,<4.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
httpx = "^0.28.1"
black = "^24.2.0"
isort = "^5.13.2"
flake8 = "^7.0.0"
mypy = "^1.8.0"
pytest-asyncio = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
